Moved Thankyou page into Page folders
