+++ frontend/src/pages/ThankyouForApplying.tsx
@@ -1,6 +1,6 @@
 import { Button, Page } from "../components/index.ts";
 
-import styles from "./ThankyouForApplying.module.css";
+import styles from "../stylesheets/ThankyouForApplying.module.css";
 
 export function ThankyouForApplying() {
   return (
