+++ frontend/src/pages/index.ts
@@ -1,7 +1,6 @@
-import { ThankyouForApplying } from "../components/ThankyouForApplying.tsx";
-
 import { Apply } from "./Apply.tsx";
 import { Candidates } from "./Candidates.tsx";
 import { TestCongratulations } from "./TestCongratulations.tsx";
+import { ThankyouForApplying } from "./ThankyouForApplying.tsx";
 
 export { Candidates, Apply, TestCongratulations, ThankyouForApplying };
