+++ frontend/src/stylesheets/ThankyouForApplying.module.css
